pyproject.toml
[project]
name = "spacebot"
version = "0.3.0"
description = "Matrix bot that auto-invites users to rooms when they join a Space"
requires-python = ">=3.14"
dependencies = [
    "matrix-nio>=0.25.2",
    "aiosqlite>=0.20.0",
]

[project.scripts]
spacebot = "spacebot.__main__:main"

[tool.uv]
package = true
